Remove "_comment" field
As it turns out, the TypeScript compiler does not allow a field called
"_comment". On the other hand, comments _are_ permitted, despite the
file having the .json extension.

https://github.com/microsoft/TypeScript/issues/4987

https://bit.ly/2k0yXZB

In this case I do not think the comment is actually needed, so I have removed it.

The error from tsc is:

Could not find implementations for the following rules specified in the configuration:
    _comment
Try upgrading TSLint and/or ensuring that you have all necessary custom rules installed.
If TSLint was recently upgraded, you may have old rules configured which need to be cleaned up.